requirements.txt
aerich
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
deprecated==1.2.13
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.70.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
jose==1.0.0
loguru==0.6.0
oss2==2.0.5
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycrypto==2.6.1
pydantic==1.10.2
pyjwt==2.5.0
pymysql==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
pyyaml==6.0
redis==4.3.4
requests==2.28.1
rsa==4.9
setuptools==65.4.1
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
tomlkit==0.11.5
tortoise-orm==0.19.2
tortoise-orm[aiomysql]==0.19.2
transaction==3.0.1
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn[standard]==0.17.6
uvloop==0.17.0
watchgod==0.8.2
websockets==10.3
win32-setctime==1.1.0
wrapt==1.14.1
zope-interface==5.4.0
gunicorn==20.1.0
xmlschema==2.1.1
etcd3==0.12.0
protobuf==3.19.0
reportlab==3.6.12
casbin==1.17.5
casbin-tortoise-adapter==1.2.0
crypto==1.4.1
pycryptodome==3.17
pycryptodomex==3.17
func-timeout==4.3.5
arrow==1.2.3
slowapi==0.1.7
pycryptodome==3.17
slowapi==0.1.7
